Fix session concurrent writes so there are no more fatal crashes! Cleanup club settings - populating form values from club object change lat & long to float32 types add function to trim zeros from float32 numbers.

+++ go/schema.go
@@ -40,8 +40,8 @@ type Club struct {
 	Name      string  `bson:"schemaName"`
 	IsDefault bool    `bson:"schemaIsDefault,omitempty"`
 	Mounds    []Mound `bson:"schemaMound,omitempty"`
-	Latitude  float64 `bson:"schemaLatitude,omitempty"`
-	Longitude float64 `bson:"schemaLongitude,omitempty"`
+	Latitude  float32 `bson:"schemaLatitude,omitempty"`
+	Longitude float32 `bson:"schemaLongitude,omitempty"`
 	URL       string  `bson:"schemaURL,omitempty"`
 	Address   string  `bson:"schemaAddress,omitempty"`
 	Town      string  `bson:"schemaTown,omitempty"`
